Updated MissionPlanner to .net4.6, which fixed their connection to non SSL3 web servers.
The only issue with this is the bitmiracle  TIFF library which does not have a 4.6 candidate in nuget. Have explicitly pointed the reference at 4.0, which resides under the nuget install. You will have to do a nuget update-pagage -reinstall

New versions of NetDXF change the name of the property "location" for "postition". Have updated these references.